Fixed use of the ObserverWrapper
Had to replace the projection for the query with null to allow the CursorWrapper to initialize the object correctly.
The original limited projection basically caused the wrapper to initialize object fields from a cursor which did
not have the corresponding columns.

Added some additional safety checks for the cursor when binding to the view.

Removed redundant parentheses in the selection statement-they get added anyway.